Updating to use partition-endpoint when no region
If no region is provided, we will use the partition wide endpoint if one
is available. This will iterate through each partition until a
partition-wide endpoint is found. This accounts for the fact that the
following endpoints currently require no region and have partition wide
endpoints: importexport, cloudfront, waf, route53, s3, sts, iam

This commit also moves several region tests into functional tests.
